fix comments and update example
